moodflixapp/requirements.txt
pytest==7.4.3
pytest-cov==2.11.1
streamlit==1.27.2
google-cloud-bigquery
google-generativeai
google-cloud-aiplatform
